u: Product 도메인 모델(int)과 일치하도록 테이블의 price 컬럼 타입을 Float에서 Integer로 변경

+++ examples/sql_example/src/product/adapter/table.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, Integer, MetaData, String, Table
+from sqlalchemy import Column, Integer, MetaData, String, Table
 
 metadata = MetaData()
 
@@ -7,7 +7,7 @@ products_table = Table(
     metadata,
     Column("id", Integer, primary_key=True, autoincrement=True),
     Column("name", String(100)),
-    Column("price", Float),
+    Column("price", Integer),
     Column("created_at", String),
     Column("updated_at", String),
 )
